+++ spec/lib/initial_avatar_spec.rb
@@ -2,13 +2,29 @@ require 'spec_helper'
 
 describe Scarf::InitialAvatar do
   describe '#svg' do
-    subject { described_class.new(name, font_family: ['Source Sans Pro']) }
+    subject { described_class.new(name) }
 
     let(:name) { 'Hello World' }
 
     it 'returns an SVG with initials in it' do
-      svg = subject.svg()
-      expect(svg).to match(/<svg.+'Source Sans Pro'.+HW.+<\/svg>/)
+      expect(subject.svg).to match(/<svg.+sans-serif.+HW.+<\/svg>/)
+    end
+
+    context 'when passed a different font' do
+      subject { described_class.new(name, font_family: ['Source Sans Pro']) }
+
+      it 'returns an SVG with custom font' do
+        expect(subject.svg).to match(/<svg.+'Source Sans Pro'.+HW.+<\/svg>/)
+      end
+    end
+
+    context 'when Scarf is configured with font_family' do
+      before { Scarf.configure { |c| c.font_family = ['Times New Roman'] } }
+      after { Scarf.configure { |c| c.font_family = %w(sans-serif) } }
+
+      it 'returns an SVG with configured font in it' do
+        expect(subject.svg).to match(/<svg.+'Times New Roman'.+HW.+<\/svg>/)
+      end
     end
   end
 end